% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifydata.R
\name{classifydata}
\alias{classifydata}
\title{Reclassify variables into binary variables based on survivalROC}
\usage{
classifydata(
time,
status,
timeslicedata,
tspatientid,
predict.time,
isfill=T
)
}
\arguments{
\item{time}{Event time or censoring time for subjects}

\item{status}{Indicator of status, 1 if death or event, 0 otherwise}

\item{timeslicedata}{list object; each dataframe of the list represent data at individual time slices}

\item{tspatientid}{list object; The unique identification number corresponding to the subject at each time slice}

\item{predict.time}{Time point to predict during drawing the ROC curve}

\item{isfill}{Logical value, used to confirm whether to fill in missing data. If it is True, then fill}
}
\value{
Returns a list of the following items:

timeslicedata Risk factors after being classified

cutoff  Optimal classification threshold
}
\description{
Based on the survivalROC package, using the survival time
and survival state, calculate the optimal cutoff for the binary classification
of each variable using data at the first time slice. According to the cutoff, the variables of
different nodes are classified into two categories.
}
\details{
According to the first time node, the survival data is used to calculate the optimal classification threshold
of different risk factors. The risk factor threshold calculated at the first time node is used to calculate
the grading of the risk factors at each time node. Return the cutoff corresponding to each risk factor and
the classified timeslicedata.
}
\examples{
data("dataset")

dataset = timedivision(X2021data,"ID","Date",period = 90,left_interval = 0.5,right_interval=1.5)

time <- list()

status <- list()

tsdata <- list()

tsid <- list()

treatment <- list()

for (i in 1:10){

 data <- dataset[dataset['time_slice']==i,]

 time <- c(time,list(data['OStime_new']))

 status <- c(status,list(data['Status_of_death']))

 tsid <- c(tsid,list(data['ID']))

 c_data <- subset(data, select = c('Age','Amount of Hepatic Lesions','Largest Diameter of Hepatic Lesions (mm)','New Lesion','Vascular Invasion','Local Lymph Node Metastasis',
                                   'Distal Metastasis','Ascites','Massive Ascites','Moderate or Mild Ascites'
                                   ,'Infected with Hepatitis','ALB',"TBLT",'PT',"AFP"))

 tsdata <- c(tsdata,list(c_data))

 c_treatment <- subset(data, select = c("Resection"))

 treatment <- c(treatment,list(c_treatment))
}


# predict.time
tsdata <- classifydata(time,status,tsdata,tsid,predict.time=365*1)

}
\seealso{
survivalROC
}
