% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Performance Evaluation of Survival Path Model}
\usage{
evaluate(
survivalpath,
minnodesize
)
}
\arguments{
\item{survivalpath}{The output of the \code{survivalpath()}  function}

\item{minnodesize}{The minimal sample size of specific node for inclusion of performance evaluation.}
}
\value{
The evaluate function returns an object, which includes timeslice, Indexofmodes and Cindex
\item{timeslice}{time slice for which performance is evaluated.}
\item{Indexofnodes}{Nodes that is used for performance evaluation in the specified time slice.}
\item{Cindex}{Harrell's concordance index (C-index) value for specified time slices.}
}
\description{
According to the survival path, using Harrell's concordance index C-index
to evaluate the discriminative ability of the survival path in each specified time slice for prognosis.
}
\details{
For patients in each specific time slice, the class of survival path is regarded as a factor when computing the C-index value.
}
