% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchsubgroup.R
\name{matchsubgroup}
\alias{matchsubgroup}
\title{Re-screen the initial time slice of the subject population and data according to the conditions}
\usage{
matchsubgroup(
time,
status,
timeslicedata,
tspatientid,
varname,
varvalue
)
}
\arguments{
\item{time}{list object;Elements sorted by time node, each element is a
Dataframe object, representing event time or censoring time for subjects}

\item{status}{list object;Elements sorted by time node, each element is a
Dataframe object, representing status, 1 if death or event, 0 otherwise.}

\item{timeslicedata}{list object; Elements sorted by time node, each element is a
Dataframe object, representing risk factors for the subject}

\item{tspatientid}{list object; Elements sorted by time node, each element represents the subjects identification number}

\item{varname}{list object;The variable used to screen subjects, and the variable
needs to be a variable in timeslicedata}

\item{varvalue}{list object;Subjects whose varname variable is equal to varvalue will be selected}
}
\value{
Returns a list of the following items:
time,state,timeslicedata,tspatientid
}
\description{
The data of the subjects that meet the conditions for the first time and the data of
subsequent time slices will also be selected. According to the given screening conditions, the eligible subjects are screened
}
\details{
According to the input time, status, variables, subject ID, etc., the data of eligible
subjects is screened through specified conditions, and the subject and the variable data of the
first and subsequent time slices are re-screened. The final returned result contains four list
objects: time, state, timeslicedata, subject ID(tspatientid).
}
\examples{
varname=list('Amount of Hepatic Lesions','Largest Diameter of Hepatic Lesions (mm)') ,varvalue=list(1,1)
df <- matchsubgroup(time,status,tsdata,tspatientid,varname=
                      list('Amount of Hepatic Lesions') ,varvalue=list(1))

}
\author{
Shen Lujun and ZhangTao
}
