% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Compare and Draw the KM curve of any given node}
\usage{
plotKM(
df,
treepoints,
mytree,
risk.table=T
)
}
\arguments{
\item{df}{"data" in the return result of the survivalpath function}

\item{treepoints}{list object;Specify the node for drawing the KM curve, which is in the survival tree}

\item{mytree}{"mytree" in the return result of the survivalpath function}

\item{risk.table}{Allowed values include:
TRUE or FALSE specifying whether to show or not the risk table. Default is FALSE."absolute" or "percentage".
Shows the absolute number and the percentage of subjects at risk by time, respectively. "abs_pct" to show both absolute number and percentage.
"nrisk_cumcensor" and "nrisk_cumevents". Show the number at risk and, the cumulative number of censoring and events, respectively.}
}
\description{
According to the survival tree, draw the KM curve of the nodes on the survival tree
}
\details{
draw the KM curve
}
\examples{
data("dataset")
dataset = timedivision(X2021data,"ID","Date",period = 90,left_interval = 0.5,right_interval=0.5)

time <- list()
status <- list()
tsdata <- list()
tsid <- list()

treatment <- list()
for (i in 1:10){

 data <- dataset[dataset['time_slice']==i,]

 time <- c(time,list(data['OStime_new']))

 status <- c(status,list(data['Status_of_death']))

 tsid <- c(tsid,list(data['ID']))

 c_data <- subset(data, select = c('Age','Amount of Hepatic Lesions','Largest Diameter of Hepatic Lesions (mm)','New Lesion','Vascular Invasion','Local Lymph Node Metastasis',
                                   'Distal Metastasis','Ascites','Massive Ascites','Moderate or Mild Ascites'
                                   ,'Infected with Hepatitis','ALB',"TBLT",'PT',"AFP"))

 tsdata <- c(tsdata,list(c_data))

 c_treatment <- subset(data, select = c("Resection"))

 treatment <- c(treatment,list(c_treatment))
}

tsdata <- classifydata(time,status,tsdata,tsid,predict.time=365*1)

result <- survivalpath(time,status,tsdata[[1]],tsid,time_slices = 10,treatments = treatment,p.value=0.05,degreeofcorrelation=0.7)

mytree <- result$tree
ggtree(mytree, color="black",linetype=1,size=1.2,ladderize = T, )+
 theme_tree2() +
 geom_text2(aes(label=label),hjust=0.6, vjust=-0.6 ,size=3.0)+
 geom_text2(aes(label=paste(node,size,mytree@data$survival,mytree@data$survivalrate,sep = "/")),hjust=0.6, vjust=-1.85 ,size=3.0)+
 #geom_point2(aes(shape=isTip, color=isTip), size=mytree1@data$os/40)+
 geom_point2(aes(shape=isTip, color=isTip), size=mytree@data$size\%/\%200+1,show.legend=F)+
 #guides(color=guide_legend(title="node name/sample number/Median survival time/Survival rate")) +
 labs(size= "Nitrogen",
      x = "TimePoints",
      y = "Survival",
      subtitle = "node_name/sample number/Median survival time/Survival rate",
      title = "Survival Tree") +
 theme(legend.title=element_blank(),legend.position = c(0.1,0.9))

#plot KM curve
treepoints = c(16,23)
plotKM(result$data, treepoints,mytree,risk.table=T)

}
\seealso{
survminer
}
