% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{survivalpath}
\alias{survivalpath}
\title{survivalpath}
\usage{
survivalpath()
}
\arguments{
\item{survivaltime}{survival time
survivalresult survival result
timepotin the number of timepoint}
}
\value{
mm all result
}
\description{
build survival path
}
\details{
for survivalpath
}
\seealso{
surv
}
\author{
zs
}
