% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalpath.R
\name{survivalpath}
\alias{survivalpath}
\title{Build survival path, generate survival tree}
\usage{
survivalpath(
time,
status,
timeslicedata,
tspatientid,
time_slices,
treatments=NULL,
num_categories=2,
p.value=0.05,
minsample = 15,
degreeofcorrelation=0.7
)
}
\arguments{
\item{time}{list object;Elements sorted by time node, each element is a
Dataframe object, representing event time or censoring time for subjects}

\item{status}{list object;Elements sorted by time node, each element is a
Dataframe object, representing status, 1 if death or event, 0 otherwise.}

\item{timeslicedata}{list object; Elements sorted by time node, each element is a
Dataframe object, representing risk factors for the subject}

\item{tspatientid}{list object; Elements sorted by time node, each element is a
Dataframe object, representing the subjects’ identification number}

\item{time_slices}{null}

\item{treatments}{default NULL.It is possible to specify the intervention measures taken
by the subjects at different time points. list object}

\item{num_categories}{default 2. The maximum number of branches that each node can divide}

\item{p.value}{p.value for univariate selection; variables less than p.value are given meaning}

\item{minsample}{Minimum sample size for branching}

\item{degreeofcorrelation}{default 0.7;When the correlation between variables is greater than this value,
the variables are considered to be correlated.}
}
\value{
Returns a list of the following items:
data dataframe:contains the main risk factors and corresponding values of each
subject divided at different time points
tree survival tree of newick structure
}
\description{
build survival path, plot survival tree
}
\details{
for survivalpath
}
\examples{
data("dataset")
dataset = timedivision(X2021data,"ID","Date",period = 90,left_interval = 0.5,right_interval=1.5)

time <- list()
status <- list()
tsdata <- list()
tsid <- list()

treatment <- list()
for (i in 1:10){

 data <- dataset[dataset['timenode']==i,]

 time <- c(time,list(data['OStime_new']))

 status <- c(status,list(data['Status_new']))

 tsid <- c(tsid,list(data['ID']))

 c_data <- subset(data, select = c('Age','Amount of Hepatic Lesions','Largest Diameter of Hepatic Lesions (mm)','New Lesion','Vascular Invasion','Local Lymph Node Metastasis',
                                   'Distal Metastasis','Ascites','Massive Ascites','Moderate or Mild Ascites'
                                   ,'Infected with Hepatitis','ALB',"TBLT",'PT',"AFP"))

 tsdata <- c(tsdata,list(c_data))

 c_treatment <- subset(data, select = c("Treatment2"))

 treatment <- c(treatment,list(c_treatment))
}

tsdata <- classifydata(time,status,tsdata,tsid,cutoff=365*1)

result <- survivalpath(time,status,tsdata[[1]],tsid,time_slices = 10,treatments = treatment,p.value=0.05,degreeofcorrelation=0.7)

mytree <- result$tree
ggtree(mytree, color="black",linetype=1,size=1.2,ladderize = T, )+
 theme_tree2() +
 geom_text2(aes(label=label),hjust=0.6, vjust=-0.6 ,size=3.0)+
 geom_text2(aes(label=paste(node,size,mytree@data$survival,mytree@data$survivalrate,sep = "/")),hjust=0.6, vjust=-1.85 ,size=3.0)+
 #geom_point2(aes(shape=isTip, color=isTip), size=mytree1@data$os/40)+
 geom_point2(aes(shape=isTip, color=isTip), size=mytree@data$size\%/\%200+1,show.legend=F)+
 #guides(color=guide_legend(title="node name/sample number/Median survival time/Survival rate")) +
 labs(size= "Nitrogen",
      x = "TimePoints",
      y = "Survival",
      subtitle = "node_name/sample number/Median survival time/Survival rate",
      title = "Survival Tree") +
 theme(legend.title=element_blank(),legend.position = c(0.1,0.9))

}
\author{
Shen Lujun and ZhangTao
}
