% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedivision.R
\name{timedivision}
\alias{timedivision}
\title{According to the specified time of the original data, select the data that conforms to the periodicity,
and generate the data format that conforms to the survivalpath function}
\usage{
timedivision(
dataset,
ID,
time,
period=30,
left_interval = 0.5,
right_interval = 1.5
)
}
\arguments{
\item{dataset}{Refer to the built-in data set of this package}

\item{ID}{Subject ID corresponding to each row of data in the dataset}

\item{time}{Represents the time when each row of data in the dataset is collected}

\item{period}{Represents the interval at which subjects with the same identification number are screened}

\item{left_interval}{The data collection time is less than the difference between the node time and
left_interval*period, that is, the value is considered to belong to the time node. For data collected
multiple times at the same time node, the earliest data is selected to represent the data collected by the node.}

\item{right_interval}{The data collection time is less than the difference between right_interval*period and the
node time, that is, the value is considered to belong to the time node. For data collected
multiple times at the same time node, the earliest data is selected to represent the data collected by the node.}
}
\value{
dataframe;Data with the same ID number, sorted by consecutive time nodes
}
\description{
Perform data preprocessing for survivalpath, filter and regularize
data entries that conform to continuous time intervals according to time.
}
\details{
for time point
}
\examples{
data("dataset")
dataset = timedivision(X2021data,"ID","Date",period = 90,left_interval = 0.5,right_interval=1.5)
}
\author{
Shen Lujun and ZhangTao
}
